Make `BaseParser` an abstract class not an interface as it is providing number default implementations now
